+++ app/src/main/java/org/breezyweather/domain/weather/model/Pollen.kt
@@ -19,6 +19,7 @@ package org.breezyweather.domain.weather.model
 import android.content.Context
 import android.graphics.Color
 import androidx.annotation.ColorInt
+import androidx.core.content.ContextCompat
 import breezyweather.domain.weather.model.Pollen
 import org.breezyweather.R
 import org.breezyweather.common.source.PollenIndexSource