+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Improvements and fixes**
 - Widgets - Improve UX of custom subtitle documentation (@codewithdipesh)
+- Main screen - Show “Negligible” inside Pollen block if there is no pollen today instead of an empty block
 
 **Translations**
 - Translations updated